Simplifying connection code.
I now require IOC restart for pmacs disconnected at startup. The IOC will
spew the mimimum number of errors if a brick is missing and will not
seg fault, but will not try to restore polling etc. if it connects later.

The previous changes allowed a late connected pmac to apparently work.
However there are too many sublties regarding the order that
things get registered with the broker and when initial values are read
from the brick. Hence a late connected pmac failed many system tests.

A solution to allow this would best be done with a rewrite using
a state machine.